Centralise some shared API adapter stuff
